use separate shoppinglist array
