--- shoppinglist-ui/src/app/shared/item.model.ts
@@ -1,8 +1,0 @@
-export class Item {
-
-    constructor(
-        public name: string, 
-        public amount: number, 
-        public id: number,
-        public visible: boolean) {}
-}
