+++ shoppinglist-ui/src/app/shared/types.ts
@@ -1,0 +1,6 @@
+export interface Item {
+  name: string;
+  amount: number;
+  id: number;
+  onShoppinglist: boolean;
+}
